Extract indexing of token bridge to separate mapping file
