+++ packages/arb-bridge-eth/src/mapping.ts
@@ -222,81 +222,3 @@ export function handleNodeRejected(event: NodeRejectedEvent): void {
   entity.status = "Rejected";
   entity.save();
 }
-
-export function handleDefaultGatewayUpdated(event: DefaultGatewayUpdatedEvent): void {
-  let entity = new DefaultGatewayUpdated(
-    event.transaction.hash.toHex() + "-" + event.logIndex.toString()
-  );
-  entity.newDefaultGateway = event.params.newDefaultGateway;
-  entity.timestamp = event.block.number;
-  entity.transactionHash = event.transaction.hash;
-  entity.blockNumber = event.block.number;
-  entity.save();
-}
-
-export function handleGatewaySet(event: GatewaySetEvent): void {
-  let entity = new GatewaySet(event.transaction.hash.toHex() + "-" + event.logIndex.toString());
-  entity.l1Token = event.params.l1Token;
-  entity.gateway = event.params.gateway;
-  entity.timestamp = event.block.number;
-  entity.transactionHash = event.transaction.hash;
-  entity.blockNumber = event.block.number;
-  entity.save();
-
-  // create entities if needed and set gateway ref
-  let gateway = getOrCreateGateway(event.params.gateway, event.block.number);
-  let token = getOrCreateToken(event.params.l1Token, event.block.number);
-  token.gateway = gateway.id;
-  token.save();
-}
-
-export function handleTransferRouted(event: TransferRoutedEvent): void {
-  let entity = new TransferRouted(event.transaction.hash.toHex() + "-" + event.logIndex.toString());
-  entity.token = event.params.token;
-  entity._userFrom = event.params._userFrom;
-  entity._userTo = event.params._userTo;
-  entity.gateway = event.params.gateway;
-  entity.timestamp = event.block.number;
-  entity.transactionHash = event.transaction.hash;
-  entity.blockNumber = event.block.number;
-  entity.save();
-}
-
-export function handleTxToL2(event: TxToL2Event): void {
-  let entity = new TxToL2(event.transaction.hash.toHex() + "-" + event.logIndex.toString());
-  entity._from = event.params._from;
-  entity._to = event.params._to;
-  entity._seqNum = event.params._seqNum;
-  entity._data = event.params._data;
-  entity.save();
-}
-
-export function handleWhitelistSourceUpdated(event: WhitelistSourceUpdatedEvent): void {
-  let entity = new WhitelistSourceUpdated(
-    event.transaction.hash.toHex() + "-" + event.logIndex.toString()
-  );
-  entity.newSource = event.params.newSource;
-  entity.timestamp = event.block.number;
-  entity.transactionHash = event.transaction.hash;
-  entity.blockNumber = event.block.number;
-  entity.save();
-}
-
-export function handleDepositInitiated(event: DepositInitiated): void {
-  let tokenDepositId =
-    event.transaction.hash.toHexString() + "-" + event.transaction.index.toString();
-
-  let tokenDeposit = new TokenDeposit(tokenDepositId);
-  tokenDeposit.amount = event.params._amount;
-  tokenDeposit.from = event.params._from;
-  tokenDeposit.to = event.params._to;
-  tokenDeposit.sequenceNumber = event.params._sequenceNumber;
-  tokenDeposit.l1Token = getOrCreateToken(event.params.l1Token, event.block.number).id;
-  tokenDeposit.isClassic = event.block.number.le(
-    BigInt.fromI32(BLOCK_OF_LAST_CLASSIC_TOKEN_DEPOSIT)
-  );
-  tokenDeposit.timestamp = event.block.timestamp;
-  tokenDeposit.transactionHash = event.transaction.hash;
-  tokenDeposit.blockCreatedAt = event.block.number;
-  tokenDeposit.save();
-}
